refactor: remove HLD to publish the repo (#12)
What does this Pull Request accomplish?
This pull request removes the HLD and its related files out of the repo.

Why should this Pull Request be merged?
As part of publicizing the repository, it is required to remove the HLD and its related files.

What testing has been done?
N.A.