feat: Update user documentation for Arbitrary session management (#24)
What does this Pull Request accomplish?
Updates the user facing document with more relevant example in terms of custom instrument which does the device communication operations.

- Update user reference guide
- Update contributing.md file to accommodate the new changes
- Update teststand sequence readme file to make the instructions simpler

Why should this Pull Request be merged?
To help the users on how to integrate any of their arbitrary sessions for custom instrument with the measurement plug-in framework.

What testing has been done?
NA